+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Only workload clusters release >= v19.1.0 can enable logging.
+- each cluster has a dedicated user
+- each cluster sends data as a different tenant
 
 ## [0.0.7] - 2023-10-03