+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2023-10-17
+
 ### Added
 
 - Add support for observability-bundle 1.0.0.
@@ -88,7 +90,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add '--vintage' toggle
 - Add controller for Workload Management Cluster using cluster.x-k8s.io/v1beta1```
 
-[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/giantswarm/logging-operator/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/giantswarm/logging-operator/compare/v0.0.7...v0.1.0
 [0.0.7]: https://github.com/giantswarm/logging-operator/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/giantswarm/logging-operator/compare/v0.0.5...v0.0.6